requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.38.0
google-cloud-storage